Fix: add some license for every files
- extract UrlParam to single file
- fix github ci error
