+++ dubbo/src/invoker/clone_body.rs
@@ -1,3 +1,19 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 use std::{
     collections::VecDeque,
     pin::Pin,
@@ -8,12 +24,11 @@ use std::{
 use bytes::{Buf, BufMut, Bytes, BytesMut};
 use futures_core::ready;
 
+use dubbo_base::StdError;
 use http::HeaderMap;
 use http_body::Body;
 use pin_project::pin_project;
 use thiserror::Error;
-
-use crate::StdError;
 
 #[derive(Error, Debug)]
 #[error("buffered body reach max capacity.")]
