+++ dubbo/src/lib.rs
@@ -39,7 +39,6 @@ use std::{future::Future, pin::Pin};
 
 pub use framework::Dubbo;
 
-pub type StdError = Box<dyn std::error::Error + Send + Sync + 'static>;
 pub type BoxFuture<T, E> = self::Pin<Box<dyn self::Future<Output = Result<T, E>> + Send + 'static>>;
 pub(crate) type Error = Box<dyn std::error::Error + Send + Sync>;
 pub type BoxBody = http_body::combinators::UnsyncBoxBody<bytes::Bytes, self::status::Status>;
