+++ dubbo/src/loadbalancer/mod.rs
@@ -1,3 +1,20 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+use dubbo_base::StdError;
 use futures_core::future::BoxFuture;
 use tower::{discover::ServiceList, ServiceExt};
 use tower_service::Service;
@@ -7,7 +24,6 @@ use crate::{
     invoker::{clone_body::CloneBody, clone_invoker::CloneInvoker},
     param::Param,
     svc::NewService,
-    StdError,
 };
 
 use crate::protocol::triple::triple_invoker::TripleInvoker;
