+++ dubbo/src/protocol/triple/triple_protocol.rs
@@ -15,11 +15,10 @@
  * limitations under the License.
  */
 
-use std::{boxed::Box, collections::HashMap};
+use std::collections::HashMap;
 
-use crate::extension::registry_extension::InterfaceName;
 use async_trait::async_trait;
-use dubbo_base::{url::UrlParam, Url};
+use dubbo_base::{registry_param::InterfaceName, url::UrlParam, Url};
 
 use super::{
     triple_exporter::TripleExporter, triple_invoker::TripleInvoker, triple_server::TripleServer,
