+++ dubbo/src/route/mod.rs
@@ -1,5 +1,23 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 use std::pin::Pin;
 
+use dubbo_base::StdError;
 use dubbo_logger::tracing::debug;
 use futures_core::{ready, Future};
 use futures_util::{future::Ready, FutureExt, TryFutureExt};
@@ -11,7 +29,6 @@ use crate::{
     invoker::clone_invoker::CloneInvoker,
     param::Param,
     svc::NewService,
-    StdError,
 };
 
 pub struct NewRoutes<N> {
@@ -20,6 +37,7 @@ pub struct NewRoutes<N> {
 
 pub struct NewRoutesFuture<S, T> {
     inner: RoutesFutureInnerState<S>,
+    #[allow(dead_code)]
     target: T,
 }
 
@@ -39,6 +57,7 @@ pub enum RoutesFutureInnerState<S> {
 
 #[derive(Clone)]
 pub struct Routes<T> {
+    #[allow(dead_code)]
     target: T,
     invokers: Vec<CloneInvoker<TripleInvoker>>,
 }
