+++ registry/zookeeper/src/lib.rs
@@ -22,18 +22,15 @@ use std::{collections::HashMap, env, sync::Arc, time::Duration};
 use async_trait::async_trait;
 use dubbo_base::{
     constants::{DUBBO_KEY, LOCALHOST_IP, PROVIDERS_KEY},
-    Url,
+    StdError, Url,
 };
 use dubbo_logger::tracing::{debug, error, info};
 use serde::{Deserialize, Serialize};
 use tokio::{select, sync::mpsc};
 use zookeeper::{Acl, CreateMode, WatchedEvent, WatchedEventType, Watcher, ZooKeeper};
 
-use dubbo::{
-    extension::registry_extension::{DiscoverStream, InterfaceName, Registry, ServiceChange},
-    StdError,
-};
-use dubbo_base::url::UrlParam;
+use dubbo::extension::registry_extension::{DiscoverStream, Registry, ServiceChange};
+use dubbo_base::{registry_param::InterfaceName, url::UrlParam};
 
 // 从url中获取服务注册的元数据
 // rawURL = fmt.Sprintf("%s://%s%s?%s", c.Protocol, host, c.Path, s)
@@ -264,23 +261,23 @@ impl Default for ZookeeperRegistry {
 impl Registry for ZookeeperRegistry {
     async fn register(&self, url: Url) -> Result<(), StdError> {
         debug!("register url: {}", url);
-        let service_name = url.query::<InterfaceName>().unwrap().value();
+        let interface_name = url.query::<InterfaceName>().unwrap().value();
         let url_str = url.as_str();
         let zk_path = format!(
             "/{}/{}/{}/{}",
-            DUBBO_KEY, service_name, PROVIDERS_KEY, url_str
+            DUBBO_KEY, interface_name, PROVIDERS_KEY, url_str
         );
         self.create_path_with_parent_check(zk_path.as_str(), LOCALHOST_IP, CreateMode::Ephemeral)?;
         Ok(())
     }
 
     async fn unregister(&self, url: Url) -> Result<(), StdError> {
-        let service_name = url.query::<InterfaceName>().unwrap().value();
+        let interface_name = url.query::<InterfaceName>().unwrap().value();
         let url_str = url.as_str();
 
         let zk_path = format!(
             "/{}/{}/{}/{}",
-            DUBBO_KEY, service_name, PROVIDERS_KEY, url_str
+            DUBBO_KEY, interface_name, PROVIDERS_KEY, url_str
         );
         self.delete_path(zk_path.as_str());
         Ok(())
@@ -288,9 +285,9 @@ impl Registry for ZookeeperRegistry {
 
     // for consumer to find the changes of providers
     async fn subscribe(&self, url: Url) -> Result<DiscoverStream, StdError> {
-        let service_name = url.query::<InterfaceName>().unwrap().value();
+        let interface_name = url.query::<InterfaceName>().unwrap().value();
 
-        let zk_path = format!("/{}/{}/{}", DUBBO_KEY, service_name, PROVIDERS_KEY);
+        let zk_path = format!("/{}/{}/{}", DUBBO_KEY, interface_name, PROVIDERS_KEY);
 
         debug!("subscribe service: {}", zk_path);
 
@@ -303,12 +300,12 @@ impl Registry for ZookeeperRegistry {
 
         let zk_client_in_task = self.zk_client.clone();
         let zk_path_in_task = zk_path.clone();
-        let service_name_in_task = service_name.clone();
+        let interface_name_in_task = interface_name.clone();
         let arc_listener_in_task = arc_listener.clone();
         tokio::spawn(async move {
             let zk_client = zk_client_in_task;
             let zk_path = zk_path_in_task;
-            let service_name = service_name_in_task;
+            let interface_name = interface_name_in_task;
             let listener = arc_listener_in_task;
 
             let mut current_urls = Vec::new();
@@ -384,9 +381,9 @@ impl Registry for ZookeeperRegistry {
     }
 
     async fn unsubscribe(&self, url: Url) -> Result<(), StdError> {
-        let service_name = url.query::<InterfaceName>().unwrap().value();
+        let interface_name = url.query::<InterfaceName>().unwrap().value();
 
-        let zk_path = format!("/{}/{}/{}", DUBBO_KEY, &service_name, PROVIDERS_KEY);
+        let zk_path = format!("/{}/{}/{}", DUBBO_KEY, &interface_name, PROVIDERS_KEY);
 
         info!("unsubscribe service: {}", zk_path);
         Ok(())
